Terraform: update to 0.6.15
AWS KMS support was only added in 0.6.13. It looks like this month nix got a
non-binary build in master, but I'd prefer to wait until that hits the next
release branch rather then trying to backport all the changes to go package
deps needed for that.

I also didn't see anything in the release notes that should effect us:

https://github.com/hashicorp/terraform/blob/v0.6.15/CHANGELOG.md